Move AutoStart select next to the idle-stop lifecycle fields
The first cut put "Auto-start when opening" on the General tab alongside
Snapshot deploy, but the related settings — Timeout, Working hours, Idle
SSH activity threshold — all live in the Runtime tab's "Idle stop" card.
Splitting the start/stop policy across two tabs hides the relationship and
forces the user to hunt for the auto-start control.

Move the SelectField into IdleStopFields so the lifecycle policy stays
together. The remote-only conditional carries over. Playwright spec
switches to the Runtime tab before asserting visibility; the test still
locks the "visible iff Remote" invariant.

Refs #331.
